refactor: extract live canvas model
